fix: 입력받은 금액으로 살 수 있는 로또 개수 계산하는 부분 파라미터 타입 int에서 double로 변경

+++ src/main/java/lotto/Calculator.java
@@ -3,8 +3,8 @@ package lotto;
 import java.util.Map;
 
 public class Calculator {
-    public int getLottoCount(int money) {
-        return money / 1000;
+    public double getLottoCount(double money) {
+        return money / 1000.0;
     }
 
     public Rank calculateRank(int correctCount, boolean bonus) {
